requirements.txt
Flask
flask-cors
transformers
torch
librosa
pydub
bcrypt
PyJWT
requests